requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.4
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
cycler==0.12.1
filelock==3.25.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
grpcio==1.78.0
hydra-core==1.3.2
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
lightning==2.6.1
lightning-utilities==0.15.3
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.1
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
pillow==12.1.1
pina-mathlab==0.2.6.post2603
propcache==0.4.1
protobuf==7.34.0
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.1
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.1
setuptools==80.10.2
six==1.17.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
torch==2.10.0
torch-geometric==2.7.0
torchmetrics==1.8.2
tqdm==4.67.3
triton==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
wheel==0.46.3
xxhash==3.6.0
yarl==1.23.0
